+++ libAudio/loadMPC.cpp
@@ -259,7 +259,7 @@ FileInfo *MPC_GetFileInfo(void *p_MPCFile)
 	ret->BitsPerSample = (p_MF->info->bitrate == 0 ? 16 : p_MF->info->bitrate);
 	ret->BitRate = p_MF->info->sample_freq;
 	ret->Channels = p_MF->info->channels;
-	ret->TotalTime = ((double)p_MF->info->samples) / ((double)ret->BitRate);
+	ret->TotalTime = p_MF->info->samples / ret->BitRate;
 
 	if (ExternalPlayback == 0)
 		p_MF->p_Playback = new Playback(ret, MPC_FillBuffer, p_MF->buffer, 8192, p_MPCFile);
