+++ libAudio/loadWAV.cpp
@@ -161,9 +161,9 @@ FileInfo *WAV_GetFileInfo(void *p_WAVFile)
 		fread(tmp, 4, 1, f_WAV);
 	}
 	fread(&p_WF->DataEnd, 4, 1, f_WAV);
-	ret->TotalTime = (float)p_WF->DataEnd / (float)ret->Channels;
-	ret->TotalTime /= ((float)ret->BitsPerSample / 8.0F);
-	ret->TotalTime /= ((float)ret->BitRate);
+	ret->TotalTime = p_WF->DataEnd / ret->Channels;
+	ret->TotalTime /= ret->BitsPerSample / 8;
+	ret->TotalTime /= ret->BitRate;
 	p_WF->DataEnd += ftell(f_WAV);
 
 	if (ExternalPlayback == 0)
